move Create*Operator to create_operator.go
